+++ fiberms/NetworkNodes.php
@@ -34,7 +34,8 @@ if ($_SERVER["REQUEST_METHOD"] == 'POST') {
 		$name = $_POST['name'];
 		$NetworkBox = $_POST['boxes'];
 		$note = $_POST['note'];
-		$OpenGIS = $_POST['OpenGIS'];
+		if ($_POST['OpenGIS'] == '') { $OpenGIS = 'NULL'; }
+		if ($_POST['OpenGIS'] != '') { $OpenGIS = $_POST['OpenGIS']; }
 		if ($_POST['SettlementGeoSpatial'] == '') { $SettlementGeoSpatial = 'NULL'; }
 		if ($_POST['SettlementGeoSpatial'] != '') { $SettlementGeoSpatial = $_POST['SettlementGeoSpatial']; }
 		if ($_POST['Building'] == '') { $building = 'NULL'; }
@@ -57,39 +58,56 @@ if ($_SERVER["REQUEST_METHOD"] == 'POST') {
 	ShowMessage($message,$error);
 } else {
     if (!isset($_GET['mode'])) {
+		require_once('backend/CableType.php');
+		
+		if (!isset($_GET['page'])) {
+			$page = 1;
+		} else {
+			$page = $_GET['page'];
+		}
+		if (isset($_GET['sort'])) {
+			$sort = $_GET['sort'];
+		} else {
+			$sort = 0;
+		}
 		$NodeId = $_GET['nodeid'];
 		if (!isset($_GET['nodeid'])) {
-			//$res = NetworkNode_SELECT($_GET['sort'],$_GET['FSort'],'');
-			$res = GetNetworkNodeList_NetworkBoxName($_GET['sort'],$_GET['FSort'],'');
+			$res = GetNetworkNodeList_NetworkBoxName($sort,$_GET['FSort'],'',$config['LinesPerPage'],($page-1)*$config['LinesPerPage']);
 		}
 		else {
     		$wr['id'] = $NodeId;
-    		//$res = NetworkNode_SELECT($_GET['sort'],$_GET['FSort'],$wr);
-			$res = GetNetworkNodeList_NetworkBoxName($_GET['sort'],$_GET['FSort'],$wr);
+			$res = GetNetworkNodeList_NetworkBoxName($sort,$_GET['FSort'],$wr,$config['LinesPerPage'],($page-1)*$config['LinesPerPage']);
 		}
 		if ($res['count'] < 1) {
 			$message = 'Узла с таким ID не существует!<br />
 			<a href="NetworkBox.php">Назад</a>';
 			ShowMessage($message,0);
 		}
+		$pages = GenPages('NetworkNodes.php?sort='.$sort.'&',ceil($res['AllPages']/$config['LinesPerPage']),$page);
 		$rows = $res['rows'];
 		$i = -1;
 		while (++$i < $res['count']) {
 			$node_arr[] = $rows[$i]['id'];
 			$node_arr[] = '<a href="NetworkNodes.php?mode=charac&nodeid='.$rows[$i]['id'].'">'.$rows[$i]['name'].'</a>';
-		    //$node_arr[] = $rows[$i]['NetworkBox'];
 			$node_arr[] = '<a href="NetworkBox.php?mode=charac&boxid='.$rows[$i]['NetworkBox'].'">'.$rows[$i]['inventoryNumber'].'</a>';
-		    $node_arr[] = $rows[$i]['note'];
 		    $node_arr[] = $rows[$i]['OpenGIS'];
 		    $node_arr[] = $rows[$i]['SettlementGeoSpatial'];
 		    $node_arr[] = $rows[$i]['Building'];
 		    $node_arr[] = $rows[$i]['Apartment'];
 			$node_arr[] = '<a href="NetworkNodes.php?mode=change&nodeid='.$rows[$i]['id'].'">Изменить</a>';
-			$node_arr[] = '<a href="NetworkNodes.php?mode=delete&nodeid='.$rows[$i]['id'].'">Удалить</a>';
+			$wr['NetworkNode'] = $rows[$i]['id'];
+			$res2 = CableLinePoint_SELECT($wr);
+			if ($res2['count'] == 0) {
+				$node_arr[] = '<a href="NetworkNodes.php?mode=delete&nodeid='.$rows[$i]['id'].'">Удалить</a>';
+			} else {
+				$node_arr[] = '';
+			}
 
 	  	}
 		$smarty->assign("data",$node_arr);
+		$smarty->assign("pages",$pages);
 	} elseif (($_GET['mode'] == 'charac') and (isset($_GET['nodeid']))) {		$smarty->assign("mode","charac");
+		require_once("backend/FS.php");
 		$NodeId = $_GET['nodeid'];
 		$res = GetNetworkNodeInfo($NodeId);
 		$rows = $res['NetworkNode']['rows'][0];
@@ -99,38 +117,48 @@ if ($_SERVER["REQUEST_METHOD"] == 'POST') {
 	  	while (++$i < $res['NetworkNode']['CableLinePoints']['count']) {
 	  		$cableline_arr[] = $ClpRows[$i]['id'];
 	  		$cableline_arr[] = $ClpRows[$i]['OpenGIS'];
-	  		$cableline_arr[] = $ClpRows[$i]['CableLine'];
+	  		$cableline_arr[] = '<a href="CableLine.php?mode=charac&cablelineid='.$ClpRows[$i]['CableLine'].'">'.$ClpRows[$i]['clname'].'</a>';
 	  		$cableline_arr[] = $ClpRows[$i]['meterSign'];
-			$cableline_arr[] = $ClpRows[$i]['NetworkNode'];
-			$cableline_arr[] = $ClpRows[$i]['note'];
-			$cableline_arr[] = $ClpRows[$i]['Apartment'];
-			$cableline_arr[] = $ClpRows[$i]['Building'];
-			$cableline_arr[] = $ClpRows[$i]['SettlementGeoSpatial'];
 			$cableline_arr[] = '<a href="CableLinePoint.php?mode=change&cablelinepointid='.$ClpRows[$i]['id'].'">Изменить</a>';
-			$cableline_arr[] = '<a href="CableLinePoint.php?mode=delete&cablelinepointid='.$ClpRows[$i]['id'].'">Удалить</a>';
+			$FiberSpliceCount = GetFiberSpliceCount($ClpRows[$i]['id']);
+			if ($FiberSpliceCount == 0) {
+				$cableline_arr[] = '<a href="CableLinePoint.php?mode=delete&cablelinepointid='.$ClpRows[$i]['id'].'">Удалить</a>';
+			} else {
+				$cableline_arr[] = '';
+			}
 	  	}
-
+		$FiberSpliceCount = GetFiberSpliceCount_NetworkNode($NodeId);
+		$ChangeDelete = '<br><a href="NetworkNodes.php?mode=change&nodeid='.$NodeId.'">Изменить</a>';
+		$wr['NetworkNode'] = $NodeId;
+		$res2 = CableLinePoint_SELECT($wr);
+		if ($res2['count'] == 0) {
+			$ChangeDelete .= '<br><a href="NetworkNodes.php?mode=delete&nodeid='.$NodeId.'">Удалить</a>';
+		}
+		
 		$smarty->assign("data",$cableline_arr);
 		$smarty->assign("id",$rows['id']);
 		$smarty->assign("name",$rows['name']);
-    	$smarty->assign("NetworkBox",$rows['NetworkBox']);
-    	$smarty->assign("note",$rows['note']);
+    	$smarty->assign("NetworkBox",$rows['inventoryNumber']);
+		$smarty->assign("FiberSpliceCount",$FiberSpliceCount);
+    	$smarty->assign("note",nl2br($rows['note']));
     	$smarty->assign("OpenGIS",$rows['OpenGIS']);
     	$smarty->assign("SettlementGeoSpatial",$rows['SettlementGeoSpatial']);
     	$smarty->assign("Building",$rows['Building']);
     	$smarty->assign("Apartment",$rows['Apartment']);
+		$smarty->assign("ChangeDelete",$ChangeDelete);
 	} elseif (($_GET['mode'] == 'change') and (isset($_GET['nodeid']))) {
 		if ($_SESSION['class'] > 1)	{
 			$message = '!!!';
 			ShowMessage($message,0);
 		}
 
-    	$smarty->assign("mode","change");
+    	$smarty->assign("mode","add_change");
+		$smarty->assign("mod","1");
 
 		$wr['id'] = $_GET['nodeid'];
     	$res = NetworkNode_SELECT(0,'',$wr);
     	if ($res['count'] < 1) {
-			$message = 'Ящика с таким ID не существует!<br />
+			$message = 'Узла с таким ID не существует!<br />
 						<a href="NetworkNodes.php">Назад</a>';
 			ShowMessage($message,0);
 		}
@@ -145,7 +173,7 @@ if ($_SERVER["REQUEST_METHOD"] == 'POST') {
     	$smarty->assign("Apartment",$rows[0]['Apartment']);
     	$NetworkBox = $rows[0]['NetworkBox'];
 
-    	$res = NetworkBox_SELECT('','');
+		$res = GetFreeNetworkBoxes($rows[0]['NetworkBox']);
 		$rows = $res['rows'];
 		$i = -1;
 		while (++$i<$res['count']) {
@@ -162,9 +190,10 @@ if ($_SERVER["REQUEST_METHOD"] == 'POST') {
 			ShowMessage($message,0);
 		}
 
-		$smarty->assign("mode","add");
+		$smarty->assign("mode","add_change");
+		$smarty->assign("mod","2");
 
-		$res = NetworkBox_SELECT('','');
+		$res = GetFreeNetworkBoxes(-1);
 		$rows = $res['rows'];
 		$i = -1;
 		while (++$i<$res['count']) {
